umgr: rework 'user' interface
In order to not expose sqlite3 to the client.

+++ c/apps/umgr/src/user.c
@@ -40,14 +40,25 @@ int user_init_database(sqlite3* db) {
   return 0;
 }
 
-int user_exists(sqlite3* db, const char* email) {
-  sqlite3_stmt* stmt;
+int user_exists(const char* email) {
+  sqlite3 *user_db;
+  sqlite3_stmt *stmt;
   int rc;
+
+  rc = db_open(user_db_fname, &user_db);
+  if (rc != SQLITE_OK) {
+    return -1;
+  }
+
+  if (user_init_database(user_db)) {
+    db_close(user_db);
+    return -1;
+  }
 
   char* sql = sqlite3_mprintf("SELECT 1 FROM user WHERE email=%Q", email);
 
-  if (sqlite3_prepare_v2(db, sql, -1, &stmt, NULL) != SQLITE_OK) {
-    fprintf(stderr, "SQLite error: %s\n", sqlite3_errmsg(db));
+  if (sqlite3_prepare_v2(user_db, sql, -1, &stmt, NULL) != SQLITE_OK) {
+    fprintf(stderr, "SQLite error: %s\n", sqlite3_errmsg(user_db));
     return -1;
   }
 
@@ -59,21 +70,34 @@ int user_exists(sqlite3* db, const char* email) {
 
   sql_stmt_free(stmt);
 
+  db_close(user_db);
+
   return rc;
 }
 
-int user_add(sqlite3* db,
-             const char* fname,
+int user_add(const char* fname,
              const char* lname,
              const char* email) {
-  sqlite3_stmt* stmt;
+  sqlite3 *user_db;
+  sqlite3_stmt *stmt;
+  int rc;
+
+  rc = db_open(user_db_fname, &user_db);
+  if (rc != SQLITE_OK) {
+    return -1;
+  }
+
+  if (user_init_database(user_db)) {
+    db_close(user_db);
+    return -1;
+  }
 
   const char *sql = "INSERT INTO user (fname, lname, email) VALUES (?1, ?2, ?3);";
 
-  if (sqlite3_prepare_v2(db, sql, -1, &stmt, NULL) != SQLITE_OK) {
+  if (sqlite3_prepare_v2(user_db, sql, -1, &stmt, NULL) != SQLITE_OK) {
     fprintf(stderr, "error preparing insert statement: %s\n",
-            sqlite3_errmsg(db));
-    db_close(db);
+            sqlite3_errmsg(user_db));
+    db_close(user_db);
     return -1;
   }
 
@@ -85,57 +109,82 @@ int user_add(sqlite3* db,
 
   if (sqlite3_step(stmt) != SQLITE_DONE) {
     fprintf(stderr, "error inserting into user table: %s\n",
-            sqlite3_errmsg(db));
+            sqlite3_errmsg(user_db));
     return -1;
   }
 
   sql_stmt_free(stmt);
 
+  db_close(user_db);
+
   return 0;
 }
 
-int user_change(sqlite3* db,
-                const char *fname,
-                const char *email) {
+int user_change(const char *fname, const char *email) {
+  sqlite3 *user_db;
   sqlite3_stmt *stmt;
-  int rv;
+  int rc;
+
+  rc = db_open(user_db_fname, &user_db);
+  if (rc != SQLITE_OK) {
+    return -1;
+  }
+
+  if (user_init_database(user_db)) {
+    db_close(user_db);
+    return -1;
+  }
 
   const char *sql = "UPDATE user SET fname=?1 WHERE email=?2;";
 
-  if (sqlite3_prepare_v2(db, sql, -1, &stmt, NULL) != SQLITE_OK) {
+  if (sqlite3_prepare_v2(user_db, sql, -1, &stmt, NULL) != SQLITE_OK) {
     fprintf(stderr, "error preparing update statement: %s\n",
-            sqlite3_errmsg(db));
+            sqlite3_errmsg(user_db));
     return -1;
   }
 
-  rv = sqlite3_bind_text(stmt, 1, fname, -1, SQLITE_STATIC);
+  rc = sqlite3_bind_text(stmt, 1, fname, -1, SQLITE_STATIC);
 
-  if (rv == SQLITE_OK)
-    rv = sqlite3_bind_text(stmt, 2, email, -1, SQLITE_STATIC);
+  if (rc == SQLITE_OK)
+    rc = sqlite3_bind_text(stmt, 2, email, -1, SQLITE_STATIC);
 
-  if (rv != SQLITE_OK) {
+  if (rc != SQLITE_OK) {
     fprintf(stderr, "error binding a value for the user table: %s\n",
-            sqlite3_errmsg(db));
+            sqlite3_errmsg(user_db));
   }
 
   if (sqlite3_step(stmt) != SQLITE_DONE) {
-    fprintf(stderr, "error updating user table: %s\n", sqlite3_errmsg(db));
+    fprintf(stderr, "error updating user table: %s\n", sqlite3_errmsg(user_db));
     return -1;
   }
 
   sql_stmt_free(stmt);
 
+  db_close(user_db);
+
   return 0;
 }
 
-int user_delete(sqlite3 *db, const char *email) {
+int user_delete(const char *email) {
+  sqlite3 *user_db;
   sqlite3_stmt *stmt;
+  int rc;
+
+  rc = db_open(user_db_fname, &user_db);
+  if (rc != SQLITE_OK) {
+    return -1;
+  }
+
+  if (user_init_database(user_db)) {
+    db_close(user_db);
+    return -1;
+  }
 
   const char *sql = "DELETE FROM user WHERE email=?1;";
 
-  if (sqlite3_prepare_v2(db, sql, -1, &stmt, NULL) != SQLITE_OK) {
+  if (sqlite3_prepare_v2(user_db, sql, -1, &stmt, NULL) != SQLITE_OK) {
     fprintf(stderr, "error preparing delete statement: %s\n",
-            sqlite3_errmsg(db));
+            sqlite3_errmsg(user_db));
     return -1;
   }
 
@@ -143,11 +192,13 @@ int user_delete(sqlite3 *db, const char *email) {
 
   if (sqlite3_step(stmt) != SQLITE_DONE) {
     fprintf(stderr, "error deleting user from table: %s\n",
-            sqlite3_errmsg(db));
+            sqlite3_errmsg(user_db));
     return -1;
   }
 
   sql_stmt_free(stmt);
+
+  db_close(user_db);
 
   return 0;
 }
